Working with dynamic path with props and useRoute function.
the router objet can't be accessed with this.$route anymore. So,
in the composition API, the url/path can be accessed by allowing the
router to pass props by adding [props: true] after the component in the
route js file, like this.

{path: '/random', component: SomeComponent, props: true }

then they can be accessed in the component generated by taking the props
in the setup(props)

To keep the routing dynamic and responsive, we can add the data in
computed or use the useRoute() function imported from vue-router.